backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
python-dotenv==1.0.1
tencentcloud-sdk-python
pydantic==2.6.4
python-multipart==0.0.9
aiofiles==23.2.1
requests==2.31.0
